requirements.txt
requests>=2.31.0
arxiv>=2.1.0
google-genai>=0.2.0

